fixes Sequelize hijacking console and adds jest verbose logging
