+++ package.json
@@ -16,7 +16,7 @@
   },
   "scripts": {
     "prepare": "husky install",
-    "test": "eslint --ext .js ./ && jest",
+    "test": "eslint --ext .js ./ && jest --verbose",
     "watch": "nodemon server",
     "start": "node server",
     "seed": "node seeds/seed.js"
